Average-over-time tracker complete
At first I had the idea to use the tweets' own timestamps and actually
record counts by timestamp in a hashtable with timestamps for keys -- so
rather than overall tweets per second/minute/hour I would have counts by
timestamp to the second and also be able to total up those lying within
the same minute and hour -- and things like, average tweets per second
_in_ any given hour or minute, that sort of thing. Then I realized that
my results would be 'inaccurate' should ever a second go by that there
were no tweets for, unlikely though that might be (but much less
unlikely if I were to run this over a stream tracking a particular
hashtag, let's say) -- because those zero-tweet seconds would not be
recorded in the data at all and I would be obliged to backfill those
somehow. Then when my head kind of hurt a little bit, I realized that
wasn't what the assignment was asking for at all anyway. It could still
be a cool/useful metric though. Might try it later. But when I realized
that my tracker would need to know start and stop times, that was part
of my realization that registering the trackers by class was kind of
problematic and the coupling was needlessly tight there, and trackers
might do well to have those #start and #stop methods. One hopes that the
times recorded are reasonably close to when the connection actually gets
made to the API but they're probably not perfect; there may be some
delay of a few seconds at the start. Another interesting problem to
think about.