Fix erroneous method name called on Twitter::URL
I was calling "extended_url" by mistake instead of "expanded_url". Specs
didn't catch this because the spec (specifically the mock tweets I was
passing) and the class were both making the same mistake. This is the
textbook case of where mocking gets you into trouble. I only noticed it
when I tried running it and got back an empty array of top domains. I'm
trying to remember what that mocking library is that checks your mocks
against the interface of the actual class of object you are mocking, but
I should look it up just to try it out here.